Add normalizer function in Dataset class #12

+++ examples/mnist/parser.py
@@ -1,6 +1,7 @@
 # internal libraries
 from dictionary import (
-    DATASET_PATH
+    DATASET_PATH,
+    TEST_SIZE
 )
 # external libraries
 import os
@@ -21,6 +22,10 @@ class Dataset():
         'testData': 'http://yann.lecun.com/exdb/mnist/t10k-images-idx3-ubyte.gz',
         'testLabels': 'http://yann.lecun.com/exdb/mnist/t10k-labels-idx1-ubyte.gz'
     }
+    xTrain = None
+    yTrain = None
+    xTest = None
+    yTest = None
     trainDataset = None
     testDataset = None
     valDataset = None
@@ -31,29 +36,32 @@ class Dataset():
         """
         pass
 
+    def normalize(self):
+        self.xTrain = self.xTrain / 255.
+        self.xTest = self.xTest / 255.
+
+
     def load(self):
         # load numpy data from file
-        trainData = np.load(DATASET_PATH + 'trainData.npy')
-        trainLabels = np.load(DATASET_PATH + 'trainLabels.npy')
-        testData = np.load(DATASET_PATH + 'testData.npy')
-        testLabels = np.load(DATASET_PATH + 'testLabels.npy')
-        # store as datasets
-        self.trainDataset = (trainData, trainLabels)
-        self.testDataset = (testData, testLabels)
+        self.xTrain = np.load(DATASET_PATH + 'trainData.npy')
+        self.yTrain = np.load(DATASET_PATH + 'trainLabels.npy')
+        self.xTest = np.load(DATASET_PATH + 'testData.npy')
+        self.yTest = np.load(DATASET_PATH + 'testLabels.npy')
 
     def get(self):
         # create path if it doesn't exist
         os.makedirs(DATASET_PATH, exist_ok=True)
         # iterate through each url
         for filename, url in self.urls.items():
-            # request compressed file
-            req = requests.get(url, allow_redirects=True)
-            open(DATASET_PATH + filename + '.gz', 'wb').write(req.content)
-            # uncompress file
-            with gzip.open(DATASET_PATH + filename + '.gz', 'rb') as source, open(DATASET_PATH + filename, 'wb') as dest:
-                shutil.copyfileobj(source, dest)
-            os.remove(DATASET_PATH + filename + '.gz')
-            # convert to npy
-            arr: np.ndarray = idx2numpy.convert_from_file(DATASET_PATH + filename)
-            np.save(DATASET_PATH + filename + '.npy', arr)
-            os.remove(DATASET_PATH + filename)
+            if not os.path.exists(DATASET_PATH + filename + '.npy'):
+                # request compressed file
+                req = requests.get(url, allow_redirects=True)
+                open(DATASET_PATH + filename + '.gz', 'wb').write(req.content)
+                # uncompress file
+                with gzip.open(DATASET_PATH + filename + '.gz', 'rb') as source, open(DATASET_PATH + filename, 'wb') as dest:
+                    shutil.copyfileobj(source, dest)
+                os.remove(DATASET_PATH + filename + '.gz')
+                # convert to npy
+                arr: np.ndarray = idx2numpy.convert_from_file(DATASET_PATH + filename)
+                np.save(DATASET_PATH + filename + '.npy', arr)
+                os.remove(DATASET_PATH + filename)
